test(agent): identity suite in its own file; ownership-guard and mixed-id pins
PR2 review round 4 (otReviewAgent):

- the GH#2273 identity-preservation suite moves to
  swm-head-identity-preservation.test.ts (the materializer file returns to
  ~612 lines; the small harness copies are deliberate - consolidating the
  shared sync fixture is queued with the post-chain test-structure pass).
  Registered in the unit config; CI shards auto-discover it.
- new regression pin: repairHeadPreservingIdentity's kaUal ownership guard
  spares another KA's operation subject referenced by a corrupted head
  (re-proves on the NEW method what the replaceHeadMetadata rows pin).
- new regression pin: a dirty head with one equivalent AND one
  non-equivalent stored id refuses preservation - the contract quantifies
  over EVERY referenced operation, and a first-match-return regression
  passes the single-id rows and fails only here. (Both rows pin properties
  the implementation has had since its first commit - regression pins, not
  fail-before rows.)
- dead descriptorHeadIdRows helper removed (unused since the selection
  started returning its own withhold plan).

Agent SWM lanes 123 green across 5 files; turbo build green.
